refactor: Update import paths from Auth to Deploy for consistency
