+++ src/redux/Store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import auth from '../screen/Auth/authSlice';
+import auth from '../screen/Deploy/authSlice';
 import home from '../screen/Home/homeSlice';
 import global from '../redux/globalSlice';
 import chat from '../screen/Chat/slices/chatSlice';
